pyproject.toml
[project]
name = "amplitude-downloader"
version = "01"
description = "0"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "n"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.117.1,<0.118.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "boto3 (>=1.40.38,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "aiohttp (>=3.12.15,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "psutil (>=7.1.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false